refactor(home): Root Providers 초기 부하 로직 protected 레이아웃으로 분리 (#445)
