+++ app/providers.tsx
@@ -8,13 +8,10 @@ import { ApiError } from "@/shared/api";
 import { AuthRouteWatcher } from "@/features/auth";
 import { AuthService, configureAuthHandlers } from "@/shared/auth";
 import { useAuthStore } from "@/entities/user";
-import { registerFcmToken } from "@/shared/firebase/registerFcmToken";
 import { registerServiceWorker } from "@/shared/pwa/registerServiceWorker";
-import { FcmForegroundListener } from "@/shared/firebase/FcmForegroundListener";
 import { useAiArrangeNoticeStore } from "@/features/home";
 import { useHomePlanStore } from "@/entities/day-plan";
 import { useUserPreferencesStore } from "@/entities/user";
-import { chatStompSession } from "@/shared/socket";
 
 export function Providers({ children }: { children: React.ReactNode }) {
   const [queryClient] = useState(
@@ -84,57 +81,10 @@ export function Providers({ children }: { children: React.ReactNode }) {
     prevAccessTokenRef.current = accessToken;
   }, [accessToken, queryClient]);
 
-  useEffect(() => {
-    if (!accessToken) return;
-    if (typeof window === "undefined") return;
-    if (!("Notification" in window)) return;
-    if (Notification.permission !== "granted") return;
-    // FCM 토큰 등록은 AppShell 탭 마운트 정책과 분리된 전역 세션 정책으로 유지한다.
-
-    let cancelled = false;
-    const run = async () => {
-      const registration = await registerServiceWorker();
-      if (!registration || cancelled) return;
-      try {
-        await registerFcmToken({ promptPermission: false });
-      } catch (error) {
-        console.error("[FCM] auto register failed", error);
-      }
-    };
-
-    void run();
-
-    return () => {
-      cancelled = true;
-    };
-  }, [accessToken]);
-
-  useEffect(() => {
-    if (!accessToken) {
-      chatStompSession.stop({
-        code: "LOGOUT",
-        message: "로그아웃으로 연결을 종료합니다.",
-      });
-      return;
-    }
-
-    chatStompSession.start(accessToken);
-  }, [accessToken]);
-
-  useEffect(() => {
-    return () => {
-      chatStompSession.stop({
-        code: "APP_UNMOUNT",
-        message: "앱 언마운트로 연결을 종료합니다.",
-      });
-    };
-  }, []);
-
   return (
     <QueryClientProvider client={queryClient}>
       <ToastProvider>
         <AuthRouteWatcher />
-        <FcmForegroundListener />
         {children}
       </ToastProvider>
       {process.env.NODE_ENV === "development" ? <ReactQueryDevtools initialIsOpen={false} /> : null}
